Fetch albums with photos for user id

+++ app/src/main/java/com/bresiu/codechallenge/data/entity/Album.java
@@ -5,13 +5,15 @@ import androidx.room.ForeignKey;
 import androidx.room.Index;
 import androidx.room.PrimaryKey;
 
-@Entity(indices = {@Index("userId")}, tableName = "albums", foreignKeys = @ForeignKey(entity = User.class, parentColumns = "id", childColumns = "userId", onDelete = ForeignKey.CASCADE))
+@Entity(indices = {
+		@Index("userId")
+}, tableName = "albums", foreignKeys = @ForeignKey(entity = User.class, parentColumns = "id", childColumns = "userId", onDelete = ForeignKey.CASCADE))
 public class Album {
-	@PrimaryKey public int id;
-	public int userId;
+	@PrimaryKey public long id;
+	public long userId;
 	public String title;
 
-	public Album(int id, int userId, String title) {
+	public Album(long id, long userId, String title) {
 		this.id = id;
 		this.userId = userId;
 		this.title = title;
