+++ app/src/main/java/com/bresiu/codechallenge/model/Album.java
@@ -1,11 +1,11 @@
 package com.bresiu.codechallenge.model;
 
 public class Album {
-	public int id;
-	public int userId;
+	public long id;
+	public long userId;
 	public String title;
 
-	public Album(int id, int userId, String title) {
+	public Album(long id, long userId, String title) {
 		this.id = id;
 		this.userId = userId;
 		this.title = title;
